Move function, delete reduant function
