+++ app/server/src/main/scala/org/bitcoins/server/Main.scala
@@ -323,32 +323,6 @@ object Main extends App with HttpLogger {
       start
     }
   }
-
-  def createWalletWithBitcoindCallbacks(
-      bitcoind: BitcoindRpcClient,
-      wallet: Wallet)(implicit ec: ExecutionContext): Wallet = {
-    // Kill the old wallet
-    wallet.stopWalletThread()
-
-    // We need to create a promise so we can inject the wallet with the callback
-    // after we have created it into SyncUtil.getNodeApiWalletCallback
-    // so we don't lose the internal state of the wallet
-    val walletCallbackP = Promise[Wallet]()
-
-    val pairedWallet = Wallet(
-      keyManager = wallet.keyManager,
-      nodeApi =
-        BitcoindRpcBackendUtil.getNodeApiWalletCallback(bitcoind,
-                                                        walletCallbackP.future),
-      chainQueryApi = BitcoindRpcBackendUtil.getChainQueryApi(bitcoind),
-      feeRateApi = wallet.feeRateApi,
-      creationTime = wallet.keyManager.creationTime
-    )(wallet.walletConfig, wallet.ec)
-
-    walletCallbackP.success(pairedWallet)
-
-    pairedWallet
-  }
 }
 
 object BitcoinSServer {
